[move] 고객 페이지 파일들 폴더 분리 #13
