+++ frontend/src/components/customer/MakePaymentScreen.js
@@ -3,8 +3,8 @@ import { useNavigate } from "react-router-dom";
 import { Box, Typography } from "@mui/material";
 import { South } from "@mui/icons-material";
 
-import { postOrder } from "../api/ApiService";
-import GoBackButton from "./button/GoBackButton";
+import { postOrder } from "../../api/ApiService";
+import GoBackButton from "../button/GoBackButton";
 import Question from "./Question";
 
 export default function MakePaymentScreen({
