+++ frontend/src/components/customer/PaymentMethodScreen.js
@@ -3,8 +3,8 @@ import { useNavigate } from "react-router-dom";
 import { Box, Grid } from "@mui/material";
 import { Money, Payment } from "@mui/icons-material";
 
-import GoBackButton from "./button/GoBackButton";
-import BoxButton from "./button/BoxButton";
+import GoBackButton from "../button/GoBackButton";
+import BoxButton from "../button/BoxButton";
 import Question from "./Question";
 
 export default function PaymentMethodScreen() {
