+++ frontend/src/components/customer/WelcomeScreen.js
